renamed extensions to jsx and moved portfolio file outside of pages directory because that directory is reserved for react components
