+++ src/containers/Blogs/Blogs.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {blogSection} from "../../pages/portfolio";
+import {blogSection} from "../../portfolio";
 import styles from "./Blogs.module.scss";
 import BlogCard from "../../components/BlogCard/BlogCard";
 import {getMediumBlogs} from "../../client/mediumBlogs";
