+++ src/containers/Skills/Skills.jsx
@@ -3,7 +3,7 @@
 import React, {useContext} from "react";
 import styles from "./Skills.module.scss";
 import SoftwareSkills from "../../components/SoftwareSkills/SoftwareSkills";
-import {illustration, skillsSection} from "../../pages/portfolio";
+import {illustration, skillsSection} from "../../portfolio";
 //import {Fade} from "react-reveal";
 import {Fade} from "react-awesome-reveal";
 import codingPerson from "../../assets/lottie/codingPerson";
